Fix XMP rotation bug
This commit addresses a bug with the rotation in the output image.
When both EXIF and XMP data is available in the header data, the
orientation set from the EXIF data will be written over once the XMP
header is read.
This makes sure that the orientation from the EXIF is preserved.